Re-wrote reporter tests using mocks to get reliable values for all kind of metrics and in the process: - Changed double[] to Double[] for percentile() and percentiles(...) (to enable mocking) - Introduced new interface Summarized which defines min(), max(), mean(), stdDev() - Introduced new interface Percentiled which defines percentile(), percentiles() and used the new interfaces to reduce duplication in reporters

+++ metrics-ganglia/src/test/java/com/yammer/metrics/reporting/GangliaReporterTest.java
@@ -148,7 +148,7 @@ public class GangliaReporterTest extends AbstractPollingReporterTest
     }
 
     @Override
-    public String[] expectedCounterResult(int count)
+    public String[] expectedCounterResult(long count)
     {
         return String.format(getFromFile("counter.io"), count).split("\\n");
     }
